feat(net,flow-entry): improve Display of flows
- Simplify the impl of Display for FlowKey so that no strings
are created for intermediate data.
- Move impl Display of FlowInfo together with impl Display for
  FlowKey and rename file.
- Adjust the layout of the flow table to improve readability
- Use upper case for protocol acronyms
