+++ net/src/flows/flow_info.rs
@@ -392,44 +392,3 @@ impl FlowInfo {
             .store(status, std::sync::atomic::Ordering::Relaxed);
     }
 }
-
-impl Display for FlowInfoLocked {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if let Some(data) = &self.dst_vpcd {
-            writeln!(f, "      dst-vpcd:{data}")?;
-        }
-        if let Some(data) = &self.port_fw_state {
-            writeln!(f, "      port-forwarding:{data}")?;
-        }
-        if let Some(data) = &self.nat_state {
-            writeln!(f, "      nat-state:{data}")?;
-        }
-        Ok(())
-    }
-}
-
-impl Display for FlowInfo {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let expires_at = self.expires_at.load(Ordering::Relaxed);
-        let expires_in = expires_at.saturating_duration_since(Instant::now());
-        let genid = self.genid();
-        writeln!(f)?;
-        if let Ok(info) = self.locked.try_read() {
-            write!(f, "{info}")?;
-        } else {
-            write!(f, "could not lock!")?;
-        }
-        let has_related = self
-            .related
-            .as_ref()
-            .and_then(std::sync::Weak::upgrade)
-            .map_or("no", |_| "yes");
-
-        writeln!(
-            f,
-            "      status: {:?}, expires in {}s, related: {has_related}, genid: {genid}",
-            self.status,
-            expires_in.as_secs(),
-        )
-    }
-}
