+++ net/src/flows/mod.rs
@@ -9,8 +9,8 @@ pub mod atomic_instant;
 pub mod flow_info;
 pub mod flow_info_item;
 
+pub mod display;
 pub mod flow_key;
-pub mod flow_key_display;
 
 pub use atomic_instant::AtomicInstant;
 pub use flow_info::*;
