+++ net/src/ip/mod.rs
@@ -122,7 +122,7 @@ impl Display for UnicastIpAddr {
 impl Display for NextHeader {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self.0.keyword_str() {
-            Some(s) => write!(f, "{}", s.to_lowercase()),
+            Some(s) => write!(f, "{s}"),
             None => write!(f, "{}", self.0.0),
         }
     }
